Rename the module names of fairseq models
