+++ knn_seq/models/__init__.py
@@ -1,5 +1,5 @@
-from .fairseq_knn_model import FairseqKNNModel
-from .fairseq_subsetknn_model import FairseqSubsetKNNModel
+from .fairseq_knn_model_vanilla import FairseqKNNModel
+from .fairseq_knn_model_subset import FairseqSubsetKNNModel
 from .hf_model import HFModelBase, build_hf_model
 
 __all__ = ["FairseqKNNModel", "FairseqSubsetKNNModel", "HFModelBase", "build_hf_model"]
