+++ knn_seq/models/fairseq_knn_model_subset.py
@@ -10,7 +10,7 @@ from torch import LongTensor, Tensor
 
 from knn_seq import utils
 from knn_seq.data.token_storage import TokenStorage
-from knn_seq.models.fairseq_knn_model import FairseqKNNModel
+from knn_seq.models.fairseq_knn_model_vanilla import FairseqKNNModel
 from knn_seq.models.hf_model import HFModelBase
 from knn_seq.search_index import FaissIndex
 from knn_seq.search_index.torch_pq_index import TorchPQIndex
